# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Overflow Outcast - Virtual Reality Art Exhibition Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /galleries:
    get:
      operationId: listGalleries
      tags: [galleries]
      summary: List authenticated artist's galleries
      responses:
        "200":
          description: List of galleries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Gallery"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createGallery
      tags: [galleries]
      summary: Create a new gallery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryInput"
      responses:
        "201":
          description: Created gallery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gallery"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /galleries/{id}:
    get:
      operationId: getGallery
      tags: [galleries]
      summary: Get a gallery by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gallery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gallery"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateGallery
      tags: [galleries]
      summary: Update a gallery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryUpdate"
      responses:
        "200":
          description: Updated gallery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gallery"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteGallery
      tags: [galleries]
      summary: Delete a gallery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /galleries/{id}/publish:
    patch:
      operationId: toggleGalleryPublish
      tags: [galleries]
      summary: Toggle gallery published state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishToggle"
      responses:
        "200":
          description: Updated gallery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gallery"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /galleries/{id}/room:
    patch:
      operationId: patchGalleryRoom
      tags: [galleries]
      summary: Update gallery room settings (theme, size, lighting, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryRoomPatch"
      responses:
        "200":
          description: Updated gallery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gallery"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /public/galleries/{slug}:
    get:
      operationId: getPublicGallery
      tags: [public]
      summary: Get a published gallery by slug with artworks
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public gallery with artworks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicGalleryDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /public/galleries/{slug}/decorations:
    get:
      operationId: getPublicGalleryDecorations
      tags: [public]
      summary: Get manual decorations for a published gallery
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of decorations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryDecoration"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Gallery:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        roomTheme:
          type: string
        published:
          type: boolean
        slug:
          type: string
        artworkCount:
          type: integer
        createdAt:
          type: string
        roomSeed:
          type: integer
        roomMode:
          type: string
        roomSize:
          type: integer
        decorationLevel:
          type: integer
        roomHeight:
          type: integer
        lightingMood:
          type: number
      required:
        - id
        - userId
        - title
        - roomTheme
        - published
        - slug
        - artworkCount
        - createdAt
        - roomSeed
        - roomMode
        - roomSize
        - decorationLevel
        - roomHeight
        - lightingMood
    GalleryInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        roomTheme:
          type: string
        roomMode:
          type: string
        roomSize:
          type: integer
          minimum: 1
          maximum: 10
        decorationLevel:
          type: integer
          minimum: 1
          maximum: 10
        roomSeed:
          type: integer
      required:
        - title
    GalleryUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        roomTheme:
          type: string
        roomMode:
          type: string
        roomSize:
          type: integer
          minimum: 1
          maximum: 10
        decorationLevel:
          type: integer
          minimum: 1
          maximum: 10
    GalleryRoomPatch:
      type: object
      properties:
        roomTheme:
          type: string
        roomSize:
          type: integer
          minimum: 1
          maximum: 10
        roomHeight:
          type: integer
          minimum: 1
          maximum: 10
        decorationLevel:
          type: integer
          minimum: 1
          maximum: 10
        lightingMood:
          type: number
          minimum: 0.5
          maximum: 2.0
        roomMode:
          type: string
    PublishToggle:
      type: object
      properties:
        published:
          type: boolean
      required:
        - published
    GalleryDecoration:
      type: object
      properties:
        id:
          type: integer
        galleryId:
          type: integer
        type:
          type: string
        x:
          type: number
        z:
          type: number
        rotY:
          type: number
        createdAt:
          type: string
      required:
        - id
        - galleryId
        - type
        - x
        - z
        - rotY
        - createdAt
...
    Artwork:
      type: object
      properties:
        id:
          type: integer
        galleryId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: string
        artistName:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
        medium:
          type: ["string", "null"]
        dimensions:
          type: ["string", "null"]
        xPosition:
          type: number
        yPosition:
          type: number
        zPosition:
          type: number
        rotation:
          type: number
        scale:
          type: number
        isManuallyPlaced:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - galleryId
        - title
        - imageUrl
        - xPosition
        - yPosition
        - zPosition
        - rotation
        - scale
        - isManuallyPlaced
        - createdAt
...
    PublicGalleryDetail:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        slug:
          type: string
        roomTheme:
          type: string
        roomSeed:
          type: integer
        roomMode:
          type: string
        roomSize:
          type: integer
        roomHeight:
          type: integer
        decorationLevel:
          type: integer
        lightingMood:
          type: number
        artistName:
          type: ["string", "null"]
        artworks:
          type: array
          items:
            $ref: "#/components/schemas/Artwork"
      required:
        - id
        - userId
        - title
        - slug
        - roomTheme
        - roomSeed
        - roomMode
        - roomSize
        - roomHeight
        - decorationLevel
        - lightingMood
        - artworks
...