# lib/api-zod/src/generated/types/gallery.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Overflow Outcast - Virtual Reality Art Exhibition Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Gallery {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  roomTheme: string;
  published: boolean;
  slug: string;
  artworkCount: number;
  createdAt: string;
  roomSeed: number;
  roomMode: string;
  roomSize: number;
  decorationLevel: number;
  roomHeight: number;
  lightingMood: number;
}