# lib/api-zod/src/generated/types/galleryRoomPatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Overflow Outcast - Virtual Reality Art Exhibition Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryRoomPatch {
  roomTheme?: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  roomSize?: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  roomHeight?: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  decorationLevel?: number;
  /**
     * @minimum 0.5
     * @maximum 2
     */
  lightingMood?: number;
  roomMode?: string;
}