# lib/api-zod/src/generated/types/publicGalleryDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Overflow Outcast - Virtual Reality Art Exhibition Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Artwork } from './artwork';

export interface PublicGalleryDetail {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  slug: string;
  roomTheme: string;
  roomSeed: number;
  roomMode: string;
  roomSize: number;
  roomHeight: number;
  decorationLevel: number;
  lightingMood: number;
  /** @nullable */
  artistName?: string | null;
  artworks: Artwork[];
}