# lib/db/src/schema/galleries.ts
import { pgTable, text, serial, integer, real, boolean, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { usersTable } from "./users";

export const galleriesTable = pgTable("galleries", {
  id: serial("id").primaryKey(),
  userId: integer("user_id").notNull().references(() => usersTable.id, { onDelete: "cascade" }),
  title: text("title").notNull(),
  description: text("description"),
  roomTheme: text("room_theme").notNull().default("dark_void"),
  published: boolean("published").notNull().default(false),
  slug: text("slug").notNull().unique(),
  roomSeed: integer("room_seed").notNull().default(0),
  roomMode: text("room_mode").notNull().default("basic"),
  roomSize: integer("room_size").notNull().default(5),
  decorationLevel: integer("decoration_level").notNull().default(5),
  roomHeight: integer("room_height").notNull().default(5),
  lightingMood: real("lighting_mood").notNull().default(1.0),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
});

export const insertGallerySchema = createInsertSchema(galleriesTable).omit({ id: true, createdAt: true });
export type InsertGallery = z.infer<typeof insertGallerySchema>;
export type Gallery = typeof galleriesTable.$inferSelect;